🔄 synced local 'skyvern/' with remote 'skyvern/'
Extracting yaml to blocks conversion as a separate step so we can do it without persisting into DB. It's needed for Copilot because LLM gives us YAML and we want to surface it to the user without persisting.

Pure refactoring, no functional changes
